// backend/src/generated/prisma/enums.ts
export const OpWorkEmploymentType = {
  FULL_TIME: 'FULL_TIME',
  PART_TIME: 'PART_TIME',
  CONTRACT: 'CONTRACT',
  INTERNSHIP: 'INTERNSHIP',
  REMOTE: 'REMOTE',
  FREELANCE: 'FREELANCE'
} as const

export type OpWorkEmploymentType = (typeof OpWorkEmploymentType)[keyof typeof OpWorkEmploymentType]
// ...
export const OpWorkEducationDegree = {
  HIGH_SCHOOL: 'HIGH_SCHOOL',
  ASSOCIATE: 'ASSOCIATE',
  BACHELOR: 'BACHELOR',
  MASTER: 'MASTER',
  DOCTORATE: 'DOCTORATE',
  CERTIFICATE: 'CERTIFICATE',
  DIPLOMA: 'DIPLOMA',
  POSTGRADUATE: 'POSTGRADUATE'
} as const

export type OpWorkEducationDegree = (typeof OpWorkEducationDegree)[keyof typeof OpWorkEducationDegree]


export const OpWorkGrade = {
  A_PLUS: 'A_PLUS',
  A: 'A',
  A_MINUS: 'A_MINUS',
  B_PLUS: 'B_PLUS',
  B: 'B',
  B_MINUS: 'B_MINUS',
  C_PLUS: 'C_PLUS',
  C: 'C',
  C_MINUS: 'C_MINUS',
  D_PLUS: 'D_PLUS',
  D: 'D',
  D_MINUS: 'D_MINUS',
  F: 'F',
  PASS: 'PASS',
  FAIL: 'FAIL',
  INCOMPLETE: 'INCOMPLETE',
  AUDIT: 'AUDIT'
} as const

export type OpWorkGrade = (typeof OpWorkGrade)[keyof typeof OpWorkGrade]